+++ spec/loan_creator/linear_spec.rb
@@ -31,5 +31,25 @@ describe LoanCreator::Linear do
         include_examples('valid lender timetable', loan_type, scenario, nil, true)
       end
     end
+
+    context 'given a scenario with term dates' do
+      term_dates = ['2022-02-15', '2022-10-15', '2023-04-15']
+
+      scenario = ['semester', '100000', '12', '2021-10-15', '3', '0']
+
+      include_examples('valid lender timetable', loan_type, scenario, {}, false, term_dates)
+    end
+
+    context 'with due interests' do
+      scenario = ['month', '1000', '12', '20202-01-01', '12', '0']
+      initial_values = {
+        paid_capital: 0,
+        paid_interests: 0,
+        accrued_delta_interests: 0.00,
+        starting_index: 1,
+        due_interests: 1000
+      }
+      include_examples('valid lender timetable', loan_type, scenario, initial_values)
+    end
   end
 end